// src/http-courses/models/empregabilidadeInformacaoComplementar.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * API Go
 * API de serviços para aplicativos da Prefeitura do Rio
 * OpenAPI spec version: 1.0
 */
import type { EmpregabilidadeTipoCampo } from './empregabilidadeTipoCampo'
import type { EmpregabilidadeVaga } from './empregabilidadeVaga'

export interface EmpregabilidadeInformacaoComplementar {
  created_at?: string
  id?: string
  id_vaga?: string
  obrigatorio?: boolean
  opcoes?: string[]
  tipo_campo?: EmpregabilidadeTipoCampo
  titulo?: string
  updated_at?: string
  /** Relationships */
  vaga?: EmpregabilidadeVaga
  valor_maximo?: number
  valor_minimo?: number
}